move <Icons> in modules
