+++ code/compiler/template.esm/src/ui/common/index.ts
@@ -1,6 +1,5 @@
-import type * as t from './t';
+export type * as t from './t';
 
-export { type t };
 export * from '../../common';
 export * from './libs';
 export * from './Icons';
