+++ code/ext/ext.lib.monaco.crdt/src/ui/common/Icons.ts
