+++ code/ext/ext.lib.monaco/src/ui/common/Icons.ts
