+++ code/ext/ext.lib.monaco/src/ui/common/index.ts
@@ -3,4 +3,4 @@ import type * as t from './t';
 export { type t };
 export * from '../../common';
 export * from './libs';
-export * from '../Icons';
+export * from './Icons';
